user.trainingreport: new module.
As part of the credentialing application, the applicant is required to
upload a PDF report from a training course.

At present, most applicants are required to take the "Data or
Specimens Only Research" course from citiprogram.org, which provides
them with a PDF file that contains a verification link - this link
allows the person reviewing the application to check that the report
is genuine.

This module defines a function 'find_training_report_url' which will
search an input file for a PDF link annotation that matches the
expected format - at present, this doesn't require any sophisticated
parsing of the document, just a regular expression search.